// web/src/types/generated/models/OrganizationRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OrganizationCategory } from './OrganizationCategory';
export type OrganizationRead = {
    id: string;
    name: string;
    bin: string;
    address: string;
    contact: string;
    email: string;
    description: string;
    category: OrganizationCategory;
    photo: (string | null);
};